[REF] Move reproducing test to `test_kfac.py`
